move used parts of couchexport.views into wrapping view
and remove couchexports.views
